[runtime] Enhancement: Move scheduler src into runtime src tree
