+++ src/rust/catnap/mod.rs
@@ -34,6 +34,10 @@ use crate::{
             Operation,
             OperationResult,
         },
+        scheduler::{
+            TaskHandle,
+            Yielder,
+        },
         types::{
             demi_opcode_t,
             demi_qresult_t,
@@ -43,10 +47,6 @@ use crate::{
         QToken,
         SharedDemiRuntime,
         SharedObject,
-    },
-    scheduler::{
-        TaskHandle,
-        Yielder,
     },
 };
 use ::std::{
