+++ src/rust/inetstack/protocols/tcp/active_open.rs
@@ -39,16 +39,16 @@ use crate::{
             types::MacAddress,
             NetworkRuntime,
         },
+        scheduler::{
+            TaskHandle,
+            Yielder,
+            YielderHandle,
+        },
         timer::SharedTimer,
         QDesc,
         SharedBox,
         SharedDemiRuntime,
         SharedObject,
-    },
-    scheduler::{
-        TaskHandle,
-        Yielder,
-        YielderHandle,
     },
 };
 use ::futures::channel::mpsc;
