+++ src/rust/runtime/scheduler/handle.rs
@@ -5,8 +5,8 @@
 // Imports
 //==============================================================================
 
-use crate::{
-    runtime::fail::Fail,
+use crate::runtime::{
+    fail::Fail,
     scheduler::page::WakerPageRef,
 };
 use ::std::{
